requirements.txt
transformers==4.40.0
torch
gradio
parler-tts
soundfile
python-dotenv
